separate_collectors_by_responsibility
