--- .circleci/config.yml
@@ -1,98 +1,0 @@
-defaults: &defaults
-  environment:
-    DOCKER_IMAGE_NAME: 31z4/ethereum-prometheus-exporter
-
-tags_filter: &tags_filter
-  tags:
-    only: /^v[0-9]+\.[0-9]+\.[0-9]+$/
-
-workflows:
-  version: 2
-  test-lint-build-push:
-    jobs:
-    - lint:
-        # See https://circleci.com/docs/2.0/workflows/#git-tag-job-execution
-        # for explanation on why we need a tags filter here.
-        filters:
-          <<: *tags_filter
-    - test:
-        filters:
-          <<: *tags_filter
-    - build:
-        filters:
-          <<: *tags_filter
-    - push:
-        requires:
-        - test
-        - lint
-        - build
-        filters:
-          <<: *tags_filter
-          branches:
-            ignore: /.*/
-
-version: 2
-jobs:
-  build:
-    <<: *defaults
-    docker:
-    - image: docker:stable-git
-    steps:
-    - checkout
-    - setup_remote_docker
-
-    - run:
-        name: Build and save a Docker image
-        command: |
-          docker build --build-arg VERSION=$(git describe --always) -t $DOCKER_IMAGE_NAME .
-          docker save --output image.tar $DOCKER_IMAGE_NAME
-    - persist_to_workspace:
-        root: .
-        paths:
-        - image.tar
-
-  test:
-    docker:
-      - image: golang:1.16
-    steps:
-      - checkout
-
-      - run:
-          name: Run tests with coverage
-          command: |
-            go test -v -race -coverprofile=coverage.txt -covermode=atomic ./...
-            bash <(curl -s https://codecov.io/bash)
-
-  lint:
-    docker:
-    - image: golang:1.16
-    steps:
-    - checkout
-
-    - run:
-        name: Download and install GolangCI-Lint
-        command: |
-          curl -sSfL https://raw.githubusercontent.com/golangci/golangci-lint/master/install.sh | sh -s -- -b $(go env GOPATH)/bin v1.41.1
-
-    - run:
-        name: Run GolangCI-Lint
-        command: |
-          golangci-lint run
-
-  push:
-    <<: *defaults
-    docker:
-    - image: docker:stable-git
-    steps:
-    - setup_remote_docker
-
-    - attach_workspace:
-        at: .
-
-    - run:
-        name: Load, tag and push a Docker image
-        command: |
-          docker load --input image.tar
-          docker tag $DOCKER_IMAGE_NAME $DOCKER_IMAGE_NAME:$CIRCLE_TAG
-          echo $DOCKER_PASSWORD | docker login -u $DOCKER_LOGIN --password-stdin
-          docker push $DOCKER_IMAGE_NAME
