--- .github/FUNDING.yml
@@ -1,2 +1,0 @@
-liberapay: 31z4
-ko_fi: elisey
