--- .golangci.yml
@@ -1,10 +1,0 @@
-linters:
-  disable-all: true
-  enable:
-  - govet
-  - errcheck
-  - megacheck
-  - ineffassign
-  - revive
-  - goimports
-  - misspell
