+++ internal/collectors/eth/gas_price.go
@@ -1,4 +1,4 @@
-package collector
+package eth
 
 import (
 	"math/big"
