+++ internal/collectors/eth/hashrate.go
@@ -1,4 +1,4 @@
-package collector
+package eth
 
 import (
 	"github.com/ethereum/go-ethereum/common/hexutil"
