+++ internal/collectors/eth/syncing.go
@@ -1,4 +1,4 @@
-package collector
+package eth
 
 import (
 	"encoding/json"
