+++ internal/collectors/eth/syncing_test.go
@@ -1,4 +1,4 @@
-package collector
+package eth
 
 import (
 	"encoding/json"
